Split the routes from main.py to separate routes based on the functionality
---

For more details, open the [Copilot Workspace session](https://copilot-workspace.githubnext.com/callingmahendra/FastAPISample?shareId=XXXX-XXXX-XXXX-XXXX).
