+++ .devcontainer.json
@@ -1,0 +1,6 @@
+{
+  "tasks": {
+    "build": "pip install -r requirements.txt",
+    "launch": "python main.py"
+  }
+}
