--- app/email.py
@@ -1,14 +1,0 @@
-import smtplib
-from email.message import EmailMessage
-
-def send_email(to_email: str, subject: str, body: str):
-    msg = EmailMessage()
-    msg.set_content(body)
-    msg['Subject'] = subject
-    msg['From'] = 'your_email@example.com'
-    msg['To'] = to_email
-
-    with smtplib.SMTP('smtp.example.com', 587) as smtp:
-        smtp.starttls()
-        smtp.login('your_email@example.com', 'your_password')
-        smtp.send_message(msg)
