+++ app/routes/customers.py
@@ -1,0 +1,47 @@
+from fastapi import APIRouter, HTTPException, Depends
+from sqlalchemy.orm import Session
+from app.database import get_db
+from app.models import Customer
+from app.schema import CustomerCreate, CustomerUpdate
+
+router = APIRouter()
+
+@router.get("/customers/")
+def get_all_customers(db: Session = Depends(get_db)):
+    return db.query(Customer).all()
+
+@router.get("/customers/{customer_id}")
+def get_customer_by_id(customer_id: int, db: Session = Depends(get_db)):
+    customer = db.query(Customer).filter(Customer.id == customer_id).first()
+    if customer:
+        return customer
+    raise HTTPException(status_code=404, detail="Customer not found")
+
+@router.post("/customers/")
+def create_customer(customer: CustomerCreate, db: Session = Depends(get_db)):
+    db_customer = Customer(name=customer.name, email=customer.email, phone=customer.phone, address=customer.address)
+    db.add(db_customer)
+    db.commit()
+    db.refresh(db_customer)
+    return db_customer
+
+@router.put("/customers/{customer_id}")
+def update_customer(customer_id: int, customer: CustomerUpdate, db: Session = Depends(get_db)):
+    db_customer = db.query(Customer).filter(Customer.id == customer_id).first()
+    if not db_customer:
+        raise HTTPException(status_code=404, detail="Customer not found")
+    db_customer.name = customer.name
+    db_customer.email = customer.email
+    db_customer.phone = customer.phone
+    db_customer.address = customer.address
+    db.commit()
+    return {"message": "Customer updated successfully"}
+
+@router.delete("/customers/{customer_id}")
+def delete_customer(customer_id: int, db: Session = Depends(get_db)):
+    db_customer = db.query(Customer).filter(Customer.id == customer_id).first()
+    if not db_customer:
+        raise HTTPException(status_code=404, detail="Customer not found")
+    db.delete(db_customer)
+    db.commit()
+    return {"message": "Customer deleted successfully"}
