+++ main.py
@@ -1,89 +1,8 @@
-from fastapi import FastAPI, HTTPException, Depends
-from sqlalchemy.orm import Session
-from app.database import get_db
-
-from app.models import User, Customer
-from app.schema import UserCreate, UserUpdate, CustomerCreate, CustomerUpdate
-from app.email import send_email  # P4eb2
-
+from fastapi import FastAPI
+from app.routes.users import router as users_router
+from app.routes.customers import router as customers_router
 
 app = FastAPI()
 
-@app.get("/users/")
-def get_all_users(db: Session = Depends(get_db)):
-    return db.query(User).all()
-
-@app.get("/users/{user_id}")
-def get_user_by_user_id(user_id: int, db: Session = Depends(get_db)):
-    user = db.query(User).filter(User.id == user_id).first()
-    if user:
-        return user
-    raise HTTPException(status_code=404, detail="User not found")
-
-@app.post("/users/")
-def create_user(user: UserCreate, db: Session = Depends(get_db)):
-    db_user = User(name=user.name, email=user.email, password=user.password)
-    db.add(db_user)
-    db.commit()
-    db.refresh(db_user)
-    send_email(db_user.email, "Welcome!", "Welcome to our platform!")  # P6ca3
-    return db_user
-
-@app.put("/users/{user_id}")
-def update_user_by_email(user_id: int, user: UserUpdate, db: Session = Depends(get_db)):
-    db_user = db.query(User).filter(User.id == user_id).first()
-    if not db_user:
-        raise HTTPException(status_code=404, detail="User not found")
-    db_user.name = user.name
-    db_user.email = user.email
-    db.commit()
-    return {"message": "User updated successfully"}
-
-@app.delete("/users/{user_id}")
-def delete_user_by_email(user_id: int, db: Session = Depends(get_db)):
-    db_user = db.query(User).filter(User.id == user_id).first()
-    if not db_user:
-        raise HTTPException(status_code=404, detail="User not found")
-    db.delete(db_user)
-    db.commit()
-    return {"message": "User deleted successfully"}
-
-@app.get("/customers/")
-def get_all_customers(db: Session = Depends(get_db)):
-    return db.query(Customer).all()
-
-@app.get("/customers/{customer_id}")
-def get_customer_by_id(customer_id: int, db: Session = Depends(get_db)):
-    customer = db.query(Customer).filter(Customer.id == customer_id).first()
-    if customer:
-        return customer
-    raise HTTPException(status_code=404, detail="Customer not found")
-
-@app.post("/customers/")
-def create_customer(customer: CustomerCreate, db: Session = Depends(get_db)):
-    db_customer = Customer(name=customer.name, email=customer.email, phone=customer.phone, address=customer.address)
-    db.add(db_customer)
-    db.commit()
-    db.refresh(db_customer)
-    return db_customer
-
-@app.put("/customers/{customer_id}")
-def update_customer(customer_id: int, customer: CustomerUpdate, db: Session = Depends(get_db)):
-    db_customer = db.query(Customer).filter(Customer.id == customer_id).first()
-    if not db_customer:
-        raise HTTPException(status_code=404, detail="Customer not found")
-    db_customer.name = customer.name
-    db_customer.email = customer.email
-    db_customer.phone = customer.phone
-    db_customer.address = customer.address
-    db.commit()
-    return {"message": "Customer updated successfully"}
-
-@app.delete("/customers/{customer_id}")
-def delete_customer(customer_id: int, db: Session = Depends(get_db)):
-    db_customer = db.query(Customer).filter(Customer.id == customer_id).first()
-    if not db_customer:
-        raise HTTPException(status_code=404, detail="Customer not found")
-    db.delete(db_customer)
-    db.commit()
-    return {"message": "Customer deleted successfully"}
+app.include_router(users_router, prefix="/users", tags=["users"])
+app.include_router(customers_router, prefix="/customers", tags=["customers"])
