+++ requirements.txt
@@ -4,35 +4,4 @@ charset-normalizer==3.1.0
 click==8.1.3
 databases==0.7.0
 dnspython==2.3.0
-email-validator==1.3.1
 fastapi==0.109.1
-fastapi-utils==0.2.1
-greenlet==2.0.2
-h11==0.14.0
-httpcore==0.16.3
-httptools==0.5.0
-httpx==0.23.3
-idna==3.4
-itsdangerous==2.1.2
-Jinja2==3.1.3
-MarkupSafe==2.1.2
-orjson==3.9.15
-psycopg2-binary==2.9.6
-pydantic==1.10.7
-python-dotenv==1.0.0
-python-multipart==0.0.7
-PyYAML==6.0
-requests==2.31.0
-rfc3986==1.5.0
-smtplib==3.6.8
-sniffio==1.3.0
-SQLAlchemy==1.4.47
-starlette<0.36.0
-typing==3.7.4.3
-typing_extensions==4.8.0
-ujson==5.7.0
-urllib3==1.26.18
-uvicorn==0.21.1
-uvloop==0.17.0
-watchfiles==0.19.0
-websockets==11.0.1
